requirements.txt
hug
gunicorn
sqlalchemy
psycopg2
alembic
sqlalchemy_utils
pytest
py-bcrypt